// old/src/app/models/external/order/order-list-item-dto.ts
import { ShippingAddressOrderDto } from "./shipping-address-order-dto";
import { InvoiceOrderDto } from "./invoice-order";
import { LatLngDto } from "../client/LatLngDto";
import { OrdenEntregaDto } from "./orden-entrega-dto";

export interface OrderListItemDto {
    id?: number;
    email?: string;
    number?: number;
    name?: string;
    note?: string;
    totalPrice?: string;
    subTotalPrice?: string;
    totalTax?: string;
    totalDiscounts?: string;
    taxesIncluded?: boolean;
    currency?: string;
    financialStatus?: string;
    fulfillmentStatus?: string;
    confirmed?: boolean;
    orderStatusUrl?: string;
    orderAdminUrl?: string;
    tags?: string[];
    itemsCount?: number;
    processedAtStr?: string;
    customerId?: number;
    formaPago?: string;
    formaPagoKeys?: string;
    customerFirstName?: string;
    customerLastName?: string;
    customerCompleteName?: string;
    shippingAddress?: ShippingAddressOrderDto;
    invoice?: InvoiceOrderDto;
    localidad?:string;
    puntoVentaId?:number;
    keyBodega?:string;
    shippingLatLng?: LatLngDto;
    ordenEntrega?: OrdenEntregaDto;
    saleUser?: string;
    saleUserId?: string;
    invoiceNumber?: string;
    invoiceName?: string;
    invoiceTypeDoc?: number;
    invoiceDocument?: string;
    invoiceComplement?: string;
    invoiceEmail?: string;
    invoicePhone?: string;
    invoiceCardNumber?: string;
    messages?: string[];
}